Write cleaned blocks at their own location.
Previously was writing out the cleaned blocks in a log structured manner
to the disk. However, we need to writem them out in a Read-Merge-Write
method - at their offset within the zone.
Fill up the holes with zeroes so that the writes match with the write
pointer of the zone.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>