Select cache zone on the number of data zones
During FG GC, we want to select a cache zone with the least number of
data zones. To enable this, we maintain a list of data zones for every
cache zone. The Cache Zone information table maintains additionally the
total number of data zones with data in the cache zone. As a block gets
added or removed, the list is checked and updated if necessary.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>