Fixed the cache overwrites
Here is a race:

	T1		T2
	overwrite	start GC
			remove translation entry (PBA, LBA)
	write completes
	(but now no translation entry exists)

This is the case of a lost write.
To prevent this we mark a flag for a cached dzone indicating that GC is
ongoing and working with data blocks belonging to this dzone.
So, whenever this flag is marked, ls_cache_write will now not perform an
overwrite, but perform a log structured write, so that when the write
completes it inserts a new translation entry in the cache. Putting it at
a new log location ensures that eventually some cache segments are
freed.

This patch also - implements a cache data zone eviction heuristic -
where a data zone is selected for eviction based on the number of data
blocks it has in the cache. The more data blocks there are, the more
attractive that cache zone becomes.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>