Cost Benefit heuristic for the data zone.
While is pure log structured architecture, cost benefit selection gives
more weight to a segment with maximum invalid blocks and oldest age,
while selecting a data zone to evict, we want to select a data zone with
the maximum number of data blocks in the cache and minimum age.
So we modified the cost benefit function for this purpose and
implemented this heuristic in this patch.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>