Wake up as many writers as released data blocks.
This patch releases as many writers as the data blocks released when a
data zone is evicted.
However - the writers does not equate the write size.
This can be worked out in an elegant manner later.
But for last minute hack - we make use of the information that fio is
always directing 1MB or 4KB writes to the device mapper.
So for the 1MB case - we unblock sectors/2048 writers
and for the 4KB case - we will unblock sectors/8 writers.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>