Re-sort the dzone list after dzone eviction.
Since the locks are taken after the dzone list is created, after the
locks are released and before they are taken, the number of blocks in
the cache that belong to a dzone changes.
So after evicting the dzone we re-sort the list as a best attempt try to
evict data zone with most dblks in the cache.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>